+++ packages/ui/src/components/SpacePanelActionCreateButton/SpacePanelActionCreateButton.tsx
@@ -30,8 +30,8 @@ export default function SpacePanelActionCreateButton({
   };
 
   if (
-    (space.tier === "free" && space.roomCount > FreeTier.roomCount) ||
-    (space.tier === "pro" && space.roomCount > ProTier.roomCount)
+    (space.tier === "free" && space.roomCount >= FreeTier.roomCount) ||
+    (space.tier === "pro" && space.roomCount >= ProTier.roomCount)
   ) {
     return (
       <>