+++ packages/web/next.config.js
@@ -1,22 +1,11 @@
-const path = require("path");
 const withPlugins = require("next-compose-plugins");
-const withCSS = require("@zeit/next-css");
-const withSass = require("@zeit/next-sass");
-// const withSourceMaps = require("@zeit/next-source-maps")();
+const withSourceMaps = require("@zeit/next-source-maps")();
 
 const withBundleAnalyzer = require("@next/bundle-analyzer")({
   enabled: process.env.ANALYZE === "true",
 });
-const {StatsWriterPlugin} = require("webpack-stats-plugin");
 
-// @ts-ignore
-const {SentryWebpackPlugin} = require("@sentry/webpack-plugin");
-
-const withBundleStats = require("next-plugin-bundle-stats")({
-  baseline: true,
-  compare: false,
-  json: true,
-});
+const SentryWebpackPlugin = require("@sentry/webpack-plugin");
 
 // @ts-ignore
 const withOptimizedImages = require("next-optimized-images")({
@@ -41,17 +30,11 @@ const withOptimizedImages = require("next-optimized-images")({
   },
 });
 
-const aliases = {
-  "@assets": path.join(__dirname, "assets"),
-  "@sentrei/common": path.join(__dirname, "../common"),
-  "@sentrei/ui": path.join(__dirname, "../ui"),
-  "@sentrei/web": path.join(__dirname, "src"),
-};
-
-const branch = String(process.env.VERCEL_GITHUB_COMMIT_REF).replace(
+const BRANCH = String(process.env.VERCEL_GITHUB_COMMIT_REF).replace(
   "refs/heads/",
   "",
 );
+
 const nextConfig = {
   target: "experimental-serverless-trace",
   trailingSlash: false,
@@ -61,6 +44,7 @@ const nextConfig = {
     SEGMENT_ID: process.env.SEGMENT_ID,
   },
   serverRuntimeConfig: {
+    rootDir: __dirname,
     FIREBASE_CLIENT_EMAIL: process.env.FIREBASE_CLIENT_EMAIL,
     FIREBASE_PRIVATE_KEY: process.env.FIREBASE_PRIVATE_KEY,
   },
@@ -80,62 +64,29 @@ const nextConfig = {
     SENTRY_DSN: process.env.SENTRY_DSN,
     SENTRY_ENVIRONMENT:
       process.env.SENTRY_ENVIRONMENT ||
-      new Set(["alpha", "beta", "main"]).has(branch)
-        ? branch
+      new Set(["alpha", "beta", "main"]).has(BRANCH)
+        ? BRANCH
         : "dev",
-    SENTRY_RELEASE: Number(require("./package.json").version),
+    SENTRY_RELEASE: require("./package.json").version,
     STRIPE_PUBLISHABLE_KEY: process.env.STRIPE_PUBLISHABLE_KEY,
-    VERCEL_GITHUB_COMMIT_REF: branch,
+    VERCEL_GITHUB_COMMIT_REF: BRANCH,
   },
   webpack: config => {
-    config.node = {
-      fs: "empty",
-      child_process: "empty",
-      net: "empty",
-      dns: "empty",
-      tls: "empty",
-    };
-    config.resolve.alias = {
-      ...config.resolve.alias,
-      ...aliases,
-    };
-    config.resolve.modules.push(path.resolve("./"));
-    if (process.env.NODE_ENV !== "production") {
-      config.plugins.push(
-        new StatsWriterPlugin({
-          filename: "webpack-stats.json",
-          stats: {
-            context: "./src",
-            assets: true,
-            entrypoints: true,
-            chunks: true,
-            modules: true,
-          },
-        }),
-      );
-    }
-    if (process.env.SENTRY_DNS) {
-      config.plugins.push(
-        new SentryWebpackPlugin({
-          include: ".next",
-          ignore: ["node_modules", "cypress", "test"],
-          urlPrefix: "~/_next",
-        }),
-      );
-    }
-    config.resolve.symlinks = true;
+    config.plugins.push(
+      // @ts-ignore
+      new SentryWebpackPlugin({
+        include: ".next",
+        ignore: ["node_modules"],
+        stripPrefix: ["webpack://_N_E/"],
+        urlPrefix: `~/_next`,
+        release: require("./package.json").version,
+      }),
+    );
     return config;
   },
 };
 
 module.exports = withPlugins(
-  [
-    [withBundleAnalyzer],
-    [withBundleStats],
-    [withCSS],
-    [withOptimizedImages],
-    [withSass],
-    // [withSourceMaps], #1462
-  ],
+  [withBundleAnalyzer, withSourceMaps, [withOptimizedImages]],
   nextConfig,
 );